feat: migrated everything to the root directory
