+++ main.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"os"
 
-	"github.com/PedroChaparro/PI202202-alako-data/scraper"
-	"github.com/PedroChaparro/PI202202-alako-data/utils"
+	"github.com/Woynert/course-mixer/scraper"
+	"github.com/Woynert/course-mixer/utils"
 )
 
 type Query struct {
