+++ samples/samples_gen.go
