+++ utils/hours.go
